Add paths-to-root functionality.
This determines a list of paths to the root concept from the specific concept. This is useful in understanding the meaning of an individual concept and potentially helpful in transforming a finely-grained concept into a more general one, particularly as this gives an ordered path to the root rather than lumping all ancestors into a cached recursive parent field.