Improved generalisation/lowest common denominator/dimensionality reduction.
This is a fairly naive implementation of a modified lowest common ancestor algorithm, in which highly granular and rarely used concepts are generalised (made more generic; ‘genericised’) to essentially reduce the dimensionality of a list of concept identifiers. It is normally better to have domain understanding and map to a refset for many analytics, thereby using a manually curated list of, say, diagnoses or problems. It may be that a future version could use manual curation falling back on naive LCA.