Remove “best match” functionality as implemented.
This is better implemented at the service level and needs extensive unit testing to be clear that it is functioning derterministically and as expected. Likewise, when we’re talking about cross maps from ICD-10, while we might want the best match that subsumes the most of its brethren, is that right? Not really, because conceptually, we need support a roundtrip, at least once, from SNOMED, to ICD-10 and back and know that the returned SNOMED at least subsumes the original concept. ie there is loss of acuity but at least ontologically, it makes sense.