Don’t yet include value sets as should be determined by number of children.
Where a root for a refinement has a limited number of children, probably defined by the client but with a server-set maximum as well, we should helpfully return the list of options as a value set, as part of that single call for refinements. However, we don’t yet cache the number of recursive children for each concept, so skip this for now.