Optimise import and precomputations.
Make import and precomputation highly parallelised. This has a very marked improvement for backend storage engines that support multiple writes; the initial store was boltdb and this had a global write lock, so multi-threading was not effective. goleveldb and online key value stores are more capable of multithreaded writes, (particularly online engines which are extraordinarily quick).