Optimise parent and child IS-A traversal
Cache IS-A relationships to give direct parents and children for any given concept. This improves the IS-A benchmark from 1,931,594 ns/op to 755,479 ns/op. Add concurrent determination of recursive children which allows the “refinements” endpoint to now include value sets, and for a new GetAllChildren endpoint, both performant. Bump database version to 2 given new indices.