Preferred synonym must be preferred AND active.
There are some descriptions which are in refsets that are labelled as preferred and yet are inactive (dm+d language refset a case in point). This is a bug, in my opinion, in the distribution, but I suppose as a “snapshot” of a changing terminology, it might be defendable that a description is labelled as preferred but inactive. Handle this by simply ignoring non-active refset items.