Add support for explicitly derived preferred terms from a single reference set, and not only language tag preferences.
This is to support the NHS dm+d realm language reference set, and the pharmacy refset,  which are separate to the standard UK language reference set, and needs to be used for dm+d / e-prescribing.

It is different because of *context*. When outside of e-prescribing, you might need “oral route” as the preferred synonym, but within e-prescribing, preferred term can be “oral” because of the context.