Use port configuration to define HTTP, and not grpc.
We have one port parameter, and previously used this for GRPC port and then ran a HTTP (REST) gateway on the next port up. This swaps these around, because most microservices expect $PORT to define a HTTP port in, say, a docker image.

Arguably, the REST gateway might be provided by the cloud provider, and we might want to explicitly define grpc port in the future, but this is fine as an interim step.