Improve mapping to a reference set.
Refine support for mapping a concept into a reference set, with appropriate specification of fallback to using parent concepts of the target reference set. This is useful for analytics so that when you want to limit your diagnoses to a simple list, you can map highly granular specialist diagnoses into something you can work with in aggregate statistics. For exampe, ADEM is a rare neuroinflammatory disorder. You can map it to the EU reference set and get “encephalitis” - perfectly appropriately when you’re trying to generate statistics or prediction models.