Improve results when maximum_hits < 100
As we re-rank search results based on length of term, if maximum hits is set very low (eg 1), then we get a very bad result. So, this defines a minimum of 100 and then discards results if they are not needed. Add unit test to be sure that we get an expected result with maximum_hits==1.